renamed api_logs to api_log
